Relocated hook from databases to src
