+++ src/hooks/useActions.ts
@@ -1,6 +1,6 @@
 import { useDispatch } from 'react-redux';
 import { bindActionCreators } from 'redux';
-import * as DataStructureActionCreators from '../../redux/DataStructure/actionCreators';
+import * as DataStructureActionCreators from '../redux/actionCreators/dataStructureActionCreators';
 
 export const useActions = () => {
   const dispatch = useDispatch();
